new network layout imple. continued

+++ BeamerLibrary/src/de/netprojectev/networking/upstream/PriorityByteEncoder.java
@@ -10,7 +10,7 @@ public class PriorityByteEncoder extends MessageToByteEncoder<Priority> {
 	@Override
 	protected void encode(ChannelHandlerContext ctx, Priority msg, ByteBuf out) throws Exception {
 		
-		out.writeByte(4); // object count
+		out.writeInt(4); // object count
 		ctx.write(msg.getId());
 		ctx.write(msg.getMinutesToShow());
 		ctx.write(msg.getName());
